remove dead code and cleanup `orderFromString`
